other.py refactored
